+++ opentelekomcloud/config.go
@@ -452,6 +452,13 @@ func (c *Config) loadEVSV2Client(region string) (*golangsdk.ServiceClient, error
 	})
 }
 
+func (c *Config) sfsV2Client(region string) (*golangsdk.ServiceClient, error) {
+	return huaweisdk.NewSharedFileSystemV2(c.HwClient, golangsdk.EndpointOpts{
+		Region:       c.determineRegion(region),
+		Availability: c.getHwEndpointType(),
+	})
+}
+
 func (c *Config) orchestrationV1Client(region string) (*golangsdk.ServiceClient, error) {
 	return huaweisdk.NewOrchestrationV1(c.HwClient, golangsdk.EndpointOpts{
 		Region:       c.determineRegion(region),